fix: shared/ui index.tsx 분리 및 각 컴포넌트 타입 설정
