+++ src/shared/ui/CardContent/CardContent.tsx
@@ -1,0 +1,10 @@
+import { forwardRef } from "react"
+
+interface ICardContentProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const CardContent = forwardRef<HTMLDivElement, ICardContentProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+))
+CardContent.displayName = "CardContent"
