+++ src/shared/ui/TableHead/TableHead.tsx
@@ -1,0 +1,15 @@
+import { forwardRef } from "react"
+
+interface ITableHeadProps extends React.HTMLAttributes<HTMLTableCellElement> {
+  className?: string
+}
+
+export const TableHead = forwardRef<HTMLTableCellElement, ITableHeadProps>(({ className, ...props }, ref) => (
+  <th
+    ref={ref}
+    className={`h-10 px-2 text-left align-middle font-medium text-muted-foreground [&:has([role=checkbox])]:pr-0 ${className}`}
+    {...props}
+  />
+))
+
+TableHead.displayName = "TableHead"
