+++ src/shared/ui/TableRow/TableRow.tsx
@@ -1,0 +1,15 @@
+import { forwardRef } from "react"
+
+interface ITableRowProps extends React.HTMLAttributes<HTMLTableRowElement> {
+  className?: string
+}
+
+export const TableRow = forwardRef<HTMLTableRowElement, ITableRowProps>(({ className, ...props }, ref) => (
+  <tr
+    ref={ref}
+    className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+    {...props}
+  />
+))
+
+TableRow.displayName = "TableRow"
