+++ src/shared/ui/index.ts
@@ -1,0 +1,15 @@
+export * from "./Button/Button"
+export * from "./Input/Input"
+export * from "./Textarea/Textarea"
+export * from "./Card/Card"
+export * from "./CardHeader/CardHeader"
+export * from "./CardTitle/CardTitle"
+export * from "./CardContent/CardContent"
+export * from "./Dialog/Dialog"
+export * from "./Select/Select"
+export * from "./Table/Table"
+export * from "./TableBody/TableBody"
+export * from "./TableCell/TableCell"
+export * from "./TableHead/TableHead"
+export * from "./TableHeader/TableHeader"
+export * from "./TableRow/TableRow"
